move api contracts to separate project that can be shared between api and sdk
