+++ backend/DevUp/DevUp.Api.Contracts/V1/Example/Requests/ExampleRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace DevUp.Api.Contracts.V1.Example.Requests
+{
+    public class ExampleRequest
+    {
+    }
+}
