+++ backend/DevUp/DevUp.Api.Contracts/V1/Example/Responses/ExampleResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace DevUp.Api.Contracts.V1.Example.Responses
+{
+    public class ExampleResponse
+    {
+    }
+}
