+++ backend/DevUp/DevUp.Api.Contracts/V1/Identity/Responses/IdentityResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace DevUp.Api.V1.Controllers.Identity.Responses
+﻿namespace DevUp.Api.Contracts.V1.Identity.Responses
 {
     public class IdentityResponse
     {
