+++ backend/DevUp/DevUp.Api/DevUp.Api.csproj
@@ -17,6 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\DevUp.Api.Contracts\DevUp.Api.Contracts.csproj" />
     <ProjectReference Include="..\DevUp.Infrastructure.Postgres\DevUp.Infrastructure.Postgres.csproj" />
     <ProjectReference Include="..\DevUp.Infrastructure\DevUp.Infrastructure.csproj" />
   </ItemGroup>
