--- backend/DevUp/DevUp.Api/V1/Controllers/Example/Requests/ExampleRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace DevUp.Api.V1.Controllers.Example.Requests
-{
-    public class ExampleRequest
-    {
-    }
-}
