--- backend/DevUp/DevUp.Api/V1/Controllers/Example/Responses/ExampleResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace DevUp.Api.V1.Controllers.Example.Responses
-{
-    public class ExampleResponse
-    {
-    }
-}
