--- backend/DevUp/DevUp.Api/V1/Controllers/Identity/Requests/RefreshUserRequest.cs
@@ -1,9 +1,0 @@
-﻿namespace DevUp.Api.V1.Controllers.Identity.Requests
-{
-    public class RefreshUserRequest
-    {
-        public string Token { get; init; }
-        public string RefreshToken { get; init; }
-        public DeviceRequest Device { get; init; }
-    }
-}
