+++ backend/DevUp/DevUp.Api/V1/Example/ExampleController.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using DevUp.Api.V1.Controllers.Example.Requests;
-using DevUp.Api.V1.Controllers.Example.Responses;
+using DevUp.Api.Contracts.V1.Example.Requests;
+using DevUp.Api.Contracts.V1.Example.Responses;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace DevUp.Api.V1.Controllers.Example
+namespace DevUp.Api.V1.Example
 {
     [Route("api/v1/[controller]")]
     [ApiController]
