+++ backend/DevUp/DevUp.Api/V1/Identity/IdentityMappingProfile.cs
@@ -1,14 +1,14 @@
 ﻿using System;
 using System.Linq;
 using AutoMapper;
-using DevUp.Api.V1.Controllers.Identity.Requests;
-using DevUp.Api.V1.Controllers.Identity.Responses;
+using DevUp.Api.Contracts.V1.Identity.Requests;
+using DevUp.Api.Contracts.V1.Identity.Responses;
 using DevUp.Domain.Identity;
 using DevUp.Domain.Identity.Entities;
 using DevUp.Domain.Identity.Exceptions;
 using DevUp.Domain.Identity.ValueObjects;
 
-namespace DevUp.Api.V1.Controllers.Identity
+namespace DevUp.Api.V1.Identity
 {
     internal class IdentityMappingProfile : Profile
     {
