+++ backend/DevUp/DevUp.sln
@@ -26,6 +26,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevUp.Tests.Utilities", "De
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DevUp.Api.Sdk", "DevUp.Api.Sdk\DevUp.Api.Sdk.csproj", "{693CAFD3-ADF1-41F1-A540-DA78AF0EC918}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DevUp.Api.Contracts", "DevUp.Api.Contracts\DevUp.Api.Contracts.csproj", "{C4CC2D18-0A3C-45FB-A1E4-148C2EFB0EB7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -64,6 +66,10 @@ Global
 		{693CAFD3-ADF1-41F1-A540-DA78AF0EC918}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{693CAFD3-ADF1-41F1-A540-DA78AF0EC918}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{693CAFD3-ADF1-41F1-A540-DA78AF0EC918}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C4CC2D18-0A3C-45FB-A1E4-148C2EFB0EB7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C4CC2D18-0A3C-45FB-A1E4-148C2EFB0EB7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C4CC2D18-0A3C-45FB-A1E4-148C2EFB0EB7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C4CC2D18-0A3C-45FB-A1E4-148C2EFB0EB7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
